refactor(host): remove containers.py back-compat shim
ContainerManager = DockerBackend alias was retained only for
E2E tests after the runtime split. Tests now import DockerBackend
directly; logger-silence list updated. No production callers.
