+++ src/cli/config.py
@@ -612,7 +612,7 @@ def _create_agent(
 def _suppress_host_logs() -> None:
     """Set host-side loggers to WARNING for clean CLI output."""
     for name in [
-        "host", "host.containers", "host.credentials",
+        "host", "host.credentials",
         "host.mesh", "host.costs", "host.permissions", "host.cron", "host.webhooks",
         "host.health", "host.lanes", "host.runtime", "host.watchers",
         "channels", "channels.base", "channels.telegram", "channels.discord",
