+++ tests/test_e2e_memory.py
@@ -62,7 +62,7 @@ def memory_stack(tmp_path_factory):
         if openai_key:
             os.environ["OPENLEGION_CRED_OPENAI_API_KEY"] = openai_key
 
-    from src.host.containers import ContainerManager
+    from src.host.runtime import DockerBackend as ContainerManager
     from src.host.credentials import CredentialVault
     from src.host.mesh import Blackboard, MessageRouter, PubSub
     from src.host.permissions import PermissionMatrix
